Pride comes before a fool. (#7)
I don't (yet) have any visual regression tests, but I relied on the
tests.  The last refactor I did knocked out all the styles and I didn't
look at the site visually until it was auto-deployed to production.

I've added a test to assert that the voting buttons are different
colours to eachother - that feels like the most basic assertion that the
styles are being included without having to name a specific colour that
should be present.  I could also assert that the title is larger than
the main text etc.

We don't have any users right now so it's not the worst thing but I'm a
bit annoyed that I was so excited about the big refactor and missed
this.